Add util script to list legends
The importer puts legends into one file as it can't tell which rows
correspond with which factions to put them in separate files (the text
in the pdf is not in the display order: it's all the profiles and then
all the headings).

It's a manual process to put this data into separate files so it can be
used by the site, but that means updates are difficult.

This util allows a ready reference of the current state of the site data
for legends so that it can be more easily manually compared to the pdf.